@brian-kelley 2024-03-26
> The bottom line is that in order to call rocSparse, the row pointers are expected to be of type `int`, which they're not right now? What's the performance difference? It might be good to put numeric results and the high level description into an accompanying issue.

That's correct.

We were already doing this for cuSPARSE CRS, so the only benefit in that case is that we don't have to size/allocate/free the temp buffer every call. This turns out to be only a few% of the cost of spmv.

But for rocSPARSE CRS, we weren't doing it so this PR now lets Tpetra use the TPL instead of the KK native. This does provide a significant speedup for the local apply:

| Time: CRS, MI250 | Before      | After       |
|------------------|-------------|-------------|
| Laplace          | 0.000419695 | 0.000117951 |
| Elasticity       | 0.000691061 | 0.000487767 |
| af_shell         | 0.000296489 | 0.000186852 |
| Hook             | 0.00100534  | 0.000614966 |

For BSR, I tried native vs. cuSPARSE 11.8 on the sparc single_gpu matrix. It turned out to be 1% slower than native (0.640328ms vs 0.64761ms).

I wanted to do a native vs. rocSPARSE BSR comparison too, but rocSPARSE doesn't have a rank-2 bsr spmv yet. It is still possible to add some logic to KokkosKernels to call the rank-1 version if there is only one column, see https://github.com/kokkos/kokkos-kernels/issues/2156.